Add fix in other classes
