+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -2,8 +2,6 @@ package ru.yandex.practicum.filmorate.model;
 
 import jakarta.validation.constraints.*;
 import lombok.Data;
-import ru.yandex.practicum.filmorate.enums.film.Genre;
-import ru.yandex.practicum.filmorate.enums.film.Rating;
 
 import java.time.LocalDate;
 import java.util.HashSet;
