+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -6,10 +6,10 @@ import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
-import ru.yandex.practicum.filmorate.storage.film.InMemoryFilmStorage;
-import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
-import ru.yandex.practicum.filmorate.storage.user.UserStorage;
+import ru.yandex.practicum.filmorate.storage.dao.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.dao.film.InMemoryFilmStorage;
+import ru.yandex.practicum.filmorate.storage.dao.user.InMemoryUserStorage;
+import ru.yandex.practicum.filmorate.storage.dao.user.UserStorage;
 
 import java.time.LocalDate;
 import java.util.*;
