+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -7,8 +7,8 @@ import ru.yandex.practicum.filmorate.exception.DuplicatedDataException;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.user.InMemoryUserStorage;
-import ru.yandex.practicum.filmorate.storage.user.UserStorage;
+import ru.yandex.practicum.filmorate.storage.dao.user.InMemoryUserStorage;
+import ru.yandex.practicum.filmorate.storage.dao.user.UserStorage;
 
 import java.time.LocalDate;
 import java.util.Collection;
