+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/film/FilmDbStorage.java
@@ -1,0 +1,108 @@
+package ru.yandex.practicum.filmorate.storage.dao.film;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.support.GeneratedKeyHolder;
+import org.springframework.jdbc.support.KeyHolder;
+import org.springframework.stereotype.Component;
+import ru.yandex.practicum.filmorate.exception.NotFoundException;
+import ru.yandex.practicum.filmorate.model.Film;
+
+import java.sql.*;
+import java.sql.Date;
+import java.util.*;
+
+@Component
+@Qualifier("SQL_Film_Storage")
+public class FilmDbStorage implements FilmStorage {
+    private final JdbcTemplate jdbcTemplate;
+
+    @Autowired
+    public FilmDbStorage(JdbcTemplate jdbcTemplate) {
+        this.jdbcTemplate = jdbcTemplate;
+    }
+
+    @Override
+    public Collection<Long> getFilmsKeys() {
+        String sql = "SELECT id FROM films";
+        return jdbcTemplate.queryForList(sql, Long.class);
+    }
+
+    @Override
+    public Collection<Film> getFilms() {
+        String sql = "SELECT * FROM films";
+        return jdbcTemplate.query(sql, this::mapRowToFilm);
+    }
+
+    @Override
+    public Film getFilm(Long id) {
+        String sql = "SELECT * FROM films WHERE id = ?";
+        List<Film> films = jdbcTemplate.query(sql, this::mapRowToFilm, id);
+        if (films.isEmpty()) {
+            throw new NotFoundException("Фильм с id = " + id + " не найден");
+        }
+        return films.get(0);
+    }
+
+    @Override
+    public Film create(Film film) {
+        String sql = "INSERT INTO films (name, description, release_date, duration, rating_id) VALUES (?, ?, ?, ?, ?)";
+        KeyHolder keyHolder = new GeneratedKeyHolder();
+        jdbcTemplate.update(connection -> {
+            PreparedStatement stmt = connection.prepareStatement(sql, new String[]{"id"});
+            stmt.setString(1, film.getName());
+            stmt.setString(2, film.getDescription());
+            stmt.setDate(3, Date.valueOf(film.getReleaseDate()));
+            stmt.setLong(4, film.getDuration());
+            stmt.setLong(5, film.getRating().ordinal() + 1);
+            return stmt;
+        }, keyHolder);
+        film.setId(Objects.requireNonNull(keyHolder.getKey()).longValue());
+        return film;
+    }
+
+    @Override
+    public Film update(Film film) {
+        String sql = "UPDATE films SET name = ?, description = ?, release_date = ?, duration = ?, rating_id = ? WHERE id = ?";
+        int updated = jdbcTemplate.update(sql,
+                film.getName(),
+                film.getDescription(),
+                Date.valueOf(film.getReleaseDate()),
+                film.getDuration(),
+                film.getRating().ordinal() + 1,
+                film.getId());
+        if (updated == 0) {
+            throw new NotFoundException("Фильм с id = " + film.getId() + " не найден");
+        }
+        return film;
+    }
+
+    @Override
+    public boolean likeFilm(Long filmId, Long userId) {
+        String sql = "INSERT INTO likes (film_id, user_id) VALUES (?, ?)";
+        try {
+            jdbcTemplate.update(sql, filmId, userId);
+            return true;
+        } catch (Exception e) {
+            return false;
+        }
+    }
+
+    @Override
+    public boolean unLikeFilm(Long filmId, Long userId) {
+        String sql = "DELETE FROM likes WHERE film_id = ? AND user_id = ?";
+        return jdbcTemplate.update(sql, filmId, userId) > 0;
+    }
+
+    private Film mapRowToFilm(ResultSet rs, int rowNum) throws SQLException {
+        Film film = new Film();
+        film.setId(rs.getLong("id"));
+        film.setName(rs.getString("name"));
+        film.setDescription(rs.getString("description"));
+        film.setReleaseDate(rs.getDate("release_date").toLocalDate());
+        film.setDuration(rs.getLong("duration"));
+        film.setRating(Rating.values()[rs.getInt("rating_id") - 1]);
+        return film;
+    }
+}
