+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/film/InMemoryFilmStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.film;
+package ru.yandex.practicum.filmorate.storage.dao.film;
 
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
