+++ src/main/java/ru/yandex/practicum/filmorate/storage/dao/user/UserStorage.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.storage.user;
+package ru.yandex.practicum.filmorate.storage.dao.user;
 
 import ru.yandex.practicum.filmorate.model.User;
