--- src/main/java/ru/yandex/practicum/filmorate/storage/film/dao/rating/FilmDbStorage.java
@@ -1,148 +1,0 @@
-package ru.yandex.practicum.filmorate.storage.film.dao.rating;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.support.GeneratedKeyHolder;
-import org.springframework.jdbc.support.KeyHolder;
-import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.enums.film.Genre;
-import ru.yandex.practicum.filmorate.enums.film.Rating;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.model.Film;
-import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
-
-import java.sql.*;
-import java.sql.Date;
-import java.util.*;
-
-@Component
-@Qualifier("SQL_Film_Storage")
-public class FilmDbStorage implements FilmStorage {
-    private final JdbcTemplate jdbcTemplate;
-
-    @Autowired
-    public FilmDbStorage(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
-
-    @Override
-    public Collection<Film> getFilms() {
-        String sql = "SELECT * FROM films";
-        return jdbcTemplate.query(sql, this::mapRowToFilm);
-    }
-
-    @Override
-    public Collection<Long> getFilmsKeys() {
-        String sql = "SELECT id FROM films";
-        return jdbcTemplate.queryForList(sql, Long.class);
-    }
-
-    @Override
-    public Film getFilm(Long id) {
-        String sql = "SELECT * FROM films WHERE id = ?";
-        List<Film> films = jdbcTemplate.query(sql, this::mapRowToFilm, id);
-        if (films.isEmpty()) {
-            throw new NotFoundException("Фильм с id = " + id + " не найден");
-        }
-        return films.get(0);
-    }
-
-    @Override
-    public boolean likeFilm(Long filmId, Long userId) {
-        String sql = "INSERT INTO likes (film_id, user_id) VALUES (?, ?)";
-        try {
-            jdbcTemplate.update(sql, filmId, userId);
-            return true;
-        } catch (Exception e) {
-            return false;
-        }
-    }
-
-    @Override
-    public boolean unLikeFilm(Long filmId, Long userId) {
-        String sql = "DELETE FROM likes WHERE film_id = ? AND user_id = ?";
-        return jdbcTemplate.update(sql, filmId, userId) > 0;
-    }
-
-    @Override
-    public Film create(Film film) {
-        String sql = "INSERT INTO films (name, description, release_date, duration, rating_id) " +
-                "VALUES (?, ?, ?, ?, ?)";
-        KeyHolder keyHolder = new GeneratedKeyHolder();
-
-        jdbcTemplate.update(connection -> {
-            PreparedStatement stmt = connection.prepareStatement(sql, new String[]{"id"});
-            stmt.setString(1, film.getName());
-            stmt.setString(2, film.getDescription());
-            stmt.setDate(3, Date.valueOf(film.getReleaseDate()));
-            stmt.setLong(4, film.getDuration());
-            stmt.setLong(5, film.getRating().ordinal() + 1);
-            return stmt;
-        }, keyHolder);
-
-        film.setId(Objects.requireNonNull(keyHolder.getKey()).longValue());
-
-        saveFilmGenres(film);
-
-        return film;
-    }
-
-    @Override
-    public Film update(Film newFilm) {
-        String sql = "UPDATE films SET name = ?, description = ?, release_date = ?, " +
-                "duration = ?, rating_id = ? WHERE id = ?";
-        int updated = jdbcTemplate.update(sql,
-                newFilm.getName(),
-                newFilm.getDescription(),
-                Date.valueOf(newFilm.getReleaseDate()),
-                newFilm.getDuration(),
-                newFilm.getRating().ordinal() + 1,
-                newFilm.getId());
-
-        if (updated == 0) {
-            throw new NotFoundException("Фильм с id = " + newFilm.getId() + " не найден");
-        }
-
-        updateFilmGenres(newFilm);
-
-        return newFilm;
-    }
-
-    private Film mapRowToFilm(ResultSet rs, int rowNum) throws SQLException {
-        Film film = new Film();
-        film.setId(rs.getLong("id"));
-        film.setName(rs.getString("name"));
-        film.setDescription(rs.getString("description"));
-        film.setReleaseDate(rs.getDate("release_date").toLocalDate());
-        film.setDuration(rs.getLong("duration"));
-        film.setRating(Rating.values()[rs.getInt("rating_id") - 1]);
-
-        String genresSql = "SELECT genre_id FROM film_genres WHERE film_id = ?";
-        List<Integer> genreIds = jdbcTemplate.queryForList(genresSql, Integer.class, film.getId());
-        if (!genreIds.isEmpty()) {
-            film.setGenre(Genre.values()[genreIds.get(0) - 1]); // Берем первый жанр
-        }
-
-        String likesSql = "SELECT user_id FROM likes WHERE film_id = ?";
-        Set<Long> likes = new HashSet<>(
-                jdbcTemplate.queryForList(likesSql, Long.class, film.getId()));
-        film.setLikesFrom(likes);
-
-        return film;
-    }
-
-    private void saveFilmGenres(Film film) {
-        if (film.getGenre() != null) {
-            String sql = "INSERT INTO film_genres (film_id, genre_id) VALUES (?, ?)";
-            jdbcTemplate.update(sql, film.getId(), film.getGenre().ordinal() + 1);
-        }
-    }
-
-    private void updateFilmGenres(Film film) {
-        String deleteSql = "DELETE FROM film_genres WHERE film_id = ?";
-        jdbcTemplate.update(deleteSql, film.getId());
-
-        saveFilmGenres(film);
-    }
-}
